corrections pour build
