+++ src/app/(main)/mesures/page.tsx
@@ -1,9 +1,9 @@
-import { getServerAuthSession } from "@/libs/server/nextAuthSession";
 import AddMeasurementForm from "./addMeasurementForm";
 import { redirect } from "next/navigation";
 import { getMeasurementsByUser } from "@/libs/server/database/measurement";
 import { Measurement } from "@/libs/types/measurement";
 import MeasurementsHistory from "./measurementsHistory";
+import { getServerAuthSession } from "@/libs/server/nextAuth";
 
 export default async function MesuresPage() {
     const session = await getServerAuthSession();
