--- src/libs/server/nextAuthSession.ts
@@ -1,13 +1,0 @@
-import { authOptions } from "@/app/api/auth/[...nextauth]/route";
-import type { GetServerSidePropsContext, NextApiRequest, NextApiResponse } from "next";
-import { getServerSession } from "next-auth";
-
-// Use it in server contexts
-export function getServerAuthSession(
-    ...args:
-        | [GetServerSidePropsContext["req"], GetServerSidePropsContext["res"]]
-        | [NextApiRequest, NextApiResponse]
-        | []
-) {
-    return getServerSession(...args, authOptions);
-}
