+++ src/libs/types/weigthGoal.ts
@@ -1,3 +1,4 @@
+/* eslint-disable no-unused-vars */
 export enum GoalStatus {
     ACTIVE = "ACTIVE",
     ARCHIVED = "ARCHIVED",
